Reorgananize all files + new toggle script for books + fix on notif and transitions
Files have been moved to folders to cleanup a bit. Toggle script to have no dependancy to the gameManager to show the 2 books (2 books can be displayed at the same time which isn't good though).

Transitions are all made by the game manager and do not depend on awake(need to fix fade in playing on awake)

notifications display correctly and only depend on the scenes that need them